chore: split accounts into separate modules
