+++ governance/program/src/entrypoint.rs
@@ -1,5 +1,4 @@
 //! Program entrypoint definitions
-
 #![cfg(all(target_arch = "bpf", not(feature = "no-entrypoint")))]
 
 use crate::{error::GovernanceError, processor};
