+++ governance/program/src/state/program_governance.rs
@@ -1,0 +1,33 @@
+//! Program Governance Account
+
+use solana_program::pubkey::Pubkey;
+
+use super::enums::GovernanceAccountType;
+
+/// Program Governance Account
+#[repr(C)]
+#[derive(Clone, Debug, Default, PartialEq)]
+pub struct ProgramGovernance {
+    /// Account type
+    pub account_type: GovernanceAccountType,
+
+    /// Voting threshold in % required to tip the vote
+    /// It's the percentage of tokens out of the entire pool of governance tokens eligible to vote
+    pub vote_threshold: u8,
+
+    /// Minimum % of tokens for a governance token owner to be able to create a proposal
+    /// It's the percentage of tokens out of the entire pool of governance tokens eligible to vote
+    pub token_threshold_to_create_proposal: u8,
+
+    /// Minimum waiting time in slots for an instruction to be executed after proposal is voted on
+    pub min_instruction_hold_up_time: u64,
+
+    /// Program ID that is governed by this Governance
+    pub program: Pubkey,
+
+    /// Time limit in slots for proposal to be open for voting
+    pub max_voting_time: u64,
+
+    /// Running count of proposals
+    pub proposal_count: u32,
+}
