+++ governance/program/src/state/single_signer_instruction.rs
@@ -1,0 +1,24 @@
+//! SingleSignerInstruction Account
+
+use solana_program::instruction::Instruction;
+
+use super::enums::GovernanceAccountType;
+
+/// Account for an instruction to be executed for Proposal
+#[repr(C)]
+#[derive(Clone)]
+pub struct SingleSignerInstruction {
+    /// Governance Account type
+    pub account_type: GovernanceAccountType,
+
+    /// Minimum waiting time in slots for the  instruction to be executed once proposal is voted on
+    pub hold_up_time: u64,
+
+    /// Instruction to execute
+    /// The instruction will be signed by Governance PDA the Proposal belongs to
+    // For example for ProgramGovernance the instruction to upgrade program will be signed by ProgramGovernance PDA
+    pub instruction: Instruction,
+
+    /// Executed flag
+    pub executed: bool,
+}
